works as intended now

+++ philosophers.h
@@ -6,7 +6,7 @@
 /*   By: stopp <stopp@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 16:08:38 by stopp             #+#    #+#             */
-/*   Updated: 2024/07/02 17:38:05 by stopp            ###   ########.fr       */
+/*   Updated: 2024/07/03 15:02:41 by stopp            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ typedef struct s_philo
 	t_mutex			*mutex;
 	struct s_philo	*next;
 	uint64_t		start_time;
-	int				curr_time;
+	uint64_t		curr_time;
 }					t_philo;
 
 /* ------------------------------- FUNCTIONS -------------------------------- */
